check if account has been verified, move logic into helper function
