+++ cartridges/app_custom_account_verification/cartridge/static/default/js/login.js
@@ -1,1 +1,1 @@
-!function(r){var e={};function t(n){if(e[n])return e[n].exports;var o=e[n]={i:n,l:!1,exports:{}};return r[n].call(o.exports,o,o.exports,t),o.l=!0,o.exports}t.m=r,t.c=e,t.d=function(r,e,n){t.o(r,e)||Object.defineProperty(r,e,{enumerable:!0,get:n})},t.r=function(r){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(r,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(r,"__esModule",{value:!0})},t.t=function(r,e){if(1&e&&(r=t(r)),8&e)return r;if(4&e&&"object"==typeof r&&r&&r.__esModule)return r;var n=Object.create(null);if(t.r(n),Object.defineProperty(n,"default",{enumerable:!0,value:r}),2&e&&"string"!=typeof r)for(var o in r)t.d(n,o,function(e){return r[e]}.bind(null,o));return n},t.n=function(r){var e=r&&r.__esModule?function(){return r.default}:function(){return r};return t.d(e,"a",e),e},t.o=function(r,e){return Object.prototype.hasOwnProperty.call(r,e)},t.p="",t(t.s=2)}([function(r,e,t){"use strict";r.exports=function(r,e){(function(r){$(r).find(".form-control.is-invalid").removeClass("is-invalid")}(r),$(".alert",r).remove(),"object"==typeof e&&e.fields&&Object.keys(e.fields).forEach((function(t){if(e.fields[t]){var n=$(r).find('[name="'+t+'"]').parent().children(".invalid-feedback");n.length>0&&(Array.isArray(e[t])?n.html(e.fields[t].join("<br/>")):n.html(e.fields[t]),n.siblings(".form-control").addClass("is-invalid"))}})),e&&e.error)&&("FORM"===$(r).prop("tagName")?$(r):$(r).parents("form")).prepend('<div class="alert alert-danger" role="alert">'+e.error.join("<br/>")+"</div>")}},function(r,e,t){"use strict";r.exports=function(r,e){var t='<div class="alert alert-danger alert-dismissible fade show" role="alert"><button type="button" class="close" data-dismiss="alert" aria-label="Close"><span aria-hidden="true">&times;</span></button>'+e+"</div>";$(r).append(t)}},function(r,e,t){"use strict";var n=t(3);$(document).ready((function(){n(t(4))}))},function(r,e,t){"use strict";r.exports=function(r){"function"==typeof r?r():"object"==typeof r&&Object.keys(r).forEach((function(e){"function"==typeof r[e]&&r[e]()}))}},function(r,e,t){"use strict";var n=t(5),o=t(0),i=t(1);n.register=function(){$("form.registration").submit((function(r){console.log("register-cusrtom");var e=$(this);r.preventDefault();var t=e.attr("action");return e.spinner().start(),$("form.registration").trigger("login:register",r),$.ajax({url:t,type:"post",dataType:"json",data:e.serialize(),success:function(r){console.log("1"),e.spinner().stop(),r.success?(console.log("login-success2"),$("form.registration").trigger("login:register:success",r),e.trigger("reset"),$(".verify-registration").removeClass("d-none"),$("#login-tab").addClass("active"),$("#login").addClass("active"),$("#register-tab").removeClass("active"),$("#register").removeClass("active")):($("form.registration").trigger("login:register:error",r),o(e,r))},error:function(r){r.responseJSON.redirectUrl?window.location.href=r.responseJSON.redirectUrl:i($(".error-messaging"),r.responseJSON.errorMessage),e.spinner().stop()}}),!1}))},r.exports=n},function(r,e,t){"use strict";var n=t(0),o=t(1);r.exports={login:function(){$("form.login").submit((function(r){var e=$(this);r.preventDefault();var t=e.attr("action");return e.spinner().start(),$("form.login").trigger("login:submit",r),$.ajax({url:t,type:"post",dataType:"json",data:e.serialize(),success:function(r){e.spinner().stop(),r.success?($("form.login").trigger("login:success",r),location.href=r.redirectUrl):(n(e,r),$("form.login").trigger("login:error",r))},error:function(r){r.responseJSON.redirectUrl?window.location.href=r.responseJSON.redirectUrl:($("form.login").trigger("login:error",r),e.spinner().stop())}}),!1}))},register:function(){$("form.registration").submit((function(r){console.log("register");var e=$(this);r.preventDefault();var t=e.attr("action");return e.spinner().start(),$("form.registration").trigger("login:register",r),$.ajax({url:t,type:"post",dataType:"json",data:e.serialize(),success:function(r){e.spinner().stop(),r.success?($("form.registration").trigger("login:register:success",r),location.href=r.redirectUrl):($("form.registration").trigger("login:register:error",r),n(e,r))},error:function(r){r.responseJSON.redirectUrl?window.location.href=r.responseJSON.redirectUrl:o($(".error-messaging"),r.responseJSON.errorMessage),e.spinner().stop()}}),!1}))},resetPassword:function(){$(".reset-password-form").submit((function(r){var e=$(this);r.preventDefault();var t=e.attr("action");return e.spinner().start(),$(".reset-password-form").trigger("login:register",r),$.ajax({url:t,type:"post",dataType:"json",data:e.serialize(),success:function(r){e.spinner().stop(),r.success?($(".request-password-title").text(r.receivedMsgHeading),$(".request-password-body").empty().append("<p>"+r.receivedMsgBody+"</p>"),r.mobile?$(".send-email-btn").empty().html('<a href="'+r.returnUrl+'" class="btn btn-primary btn-block">'+r.buttonText+"</a>"):$("#submitEmailButton").text(r.buttonText).attr("data-dismiss","modal")):n(e,r)},error:function(){e.spinner().stop()}}),!1}))},clearResetForm:function(){$("#login .modal").on("hidden.bs.modal",(function(){$("#reset-password-email").val(""),$(".modal-dialog .form-control.is-invalid").removeClass("is-invalid")}))}}}]);
+!function(e){var r={};function t(n){if(r[n])return r[n].exports;var i=r[n]={i:n,l:!1,exports:{}};return e[n].call(i.exports,i,i.exports,t),i.l=!0,i.exports}t.m=e,t.c=r,t.d=function(e,r,n){t.o(e,r)||Object.defineProperty(e,r,{enumerable:!0,get:n})},t.r=function(e){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(e,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(e,"__esModule",{value:!0})},t.t=function(e,r){if(1&r&&(e=t(e)),8&r)return e;if(4&r&&"object"==typeof e&&e&&e.__esModule)return e;var n=Object.create(null);if(t.r(n),Object.defineProperty(n,"default",{enumerable:!0,value:e}),2&r&&"string"!=typeof e)for(var i in e)t.d(n,i,function(r){return e[r]}.bind(null,i));return n},t.n=function(e){var r=e&&e.__esModule?function(){return e.default}:function(){return e};return t.d(r,"a",r),r},t.o=function(e,r){return Object.prototype.hasOwnProperty.call(e,r)},t.p="",t(t.s=2)}([function(e,r,t){"use strict";e.exports=function(e,r){(function(e){$(e).find(".form-control.is-invalid").removeClass("is-invalid")}(e),$(".alert",e).remove(),"object"==typeof r&&r.fields&&Object.keys(r.fields).forEach((function(t){if(r.fields[t]){var n=$(e).find('[name="'+t+'"]').parent().children(".invalid-feedback");n.length>0&&(Array.isArray(r[t])?n.html(r.fields[t].join("<br/>")):n.html(r.fields[t]),n.siblings(".form-control").addClass("is-invalid"))}})),r&&r.error)&&("FORM"===$(e).prop("tagName")?$(e):$(e).parents("form")).prepend('<div class="alert alert-danger" role="alert">'+r.error.join("<br/>")+"</div>")}},function(e,r,t){"use strict";e.exports=function(e,r){var t='<div class="alert alert-danger alert-dismissible fade show" role="alert"><button type="button" class="close" data-dismiss="alert" aria-label="Close"><span aria-hidden="true">&times;</span></button>'+r+"</div>";$(e).append(t)}},function(e,r,t){"use strict";var n=t(3);$(document).ready((function(){n(t(4))}))},function(e,r,t){"use strict";e.exports=function(e){"function"==typeof e?e():"object"==typeof e&&Object.keys(e).forEach((function(r){"function"==typeof e[r]&&e[r]()}))}},function(e,r,t){"use strict";var n=t(5),i=t(0),o=t(1);n.register=function(){$("form.registration").submit((function(e){console.log("register-cusrtom");var r=$(this);e.preventDefault();var t=r.attr("action");return r.spinner().start(),$("form.registration").trigger("login:register",e),$.ajax({url:t,type:"post",dataType:"json",data:r.serialize(),success:function(e){console.log("1"),r.spinner().stop(),e.success?(console.log("login-success2"),$("form.registration").trigger("login:register:success",e),r.trigger("reset"),$(".verify-registration").removeClass("d-none"),$("#login-tab").addClass("active"),$("#login").addClass("active"),$("#register-tab").removeClass("active"),$("#register").removeClass("active"),$(".verify-link-expired").addClass("d-none")):($("form.registration").trigger("login:register:error",e),i(r,e))},error:function(e){e.responseJSON.redirectUrl?window.location.href=e.responseJSON.redirectUrl:o($(".error-messaging"),e.responseJSON.errorMessage),r.spinner().stop()}}),!1}))},e.exports=n},function(e,r,t){"use strict";var n=t(0),i=t(1);e.exports={login:function(){$("form.login").submit((function(e){var r=$(this);e.preventDefault();var t=r.attr("action");return r.spinner().start(),$("form.login").trigger("login:submit",e),$.ajax({url:t,type:"post",dataType:"json",data:r.serialize(),success:function(e){r.spinner().stop(),e.success?($("form.login").trigger("login:success",e),location.href=e.redirectUrl):(n(r,e),$("form.login").trigger("login:error",e))},error:function(e){e.responseJSON.redirectUrl?window.location.href=e.responseJSON.redirectUrl:($("form.login").trigger("login:error",e),r.spinner().stop())}}),!1}))},register:function(){$("form.registration").submit((function(e){var r=$(this);e.preventDefault();var t=r.attr("action");return r.spinner().start(),$("form.registration").trigger("login:register",e),$.ajax({url:t,type:"post",dataType:"json",data:r.serialize(),success:function(e){r.spinner().stop(),e.success?($("form.registration").trigger("login:register:success",e),location.href=e.redirectUrl):($("form.registration").trigger("login:register:error",e),n(r,e))},error:function(e){e.responseJSON.redirectUrl?window.location.href=e.responseJSON.redirectUrl:i($(".error-messaging"),e.responseJSON.errorMessage),r.spinner().stop()}}),!1}))},resetPassword:function(){$(".reset-password-form").submit((function(e){var r=$(this);e.preventDefault();var t=r.attr("action");return r.spinner().start(),$(".reset-password-form").trigger("login:register",e),$.ajax({url:t,type:"post",dataType:"json",data:r.serialize(),success:function(e){r.spinner().stop(),e.success?($(".request-password-title").text(e.receivedMsgHeading),$(".request-password-body").empty().append("<p>"+e.receivedMsgBody+"</p>"),e.mobile?$(".send-email-btn").empty().html('<a href="'+e.returnUrl+'" class="btn btn-primary btn-block">'+e.buttonText+"</a>"):$("#submitEmailButton").text(e.buttonText).attr("data-dismiss","modal")):n(r,e)},error:function(){r.spinner().stop()}}),!1}))},clearResetForm:function(){$("#login .modal").on("hidden.bs.modal",(function(){$("#reset-password-email").val(""),$(".modal-dialog .form-control.is-invalid").removeClass("is-invalid")}))}}}]);
