+++ cartridges/app_custom_account_verification/cartridge/templates/resources/verify.properties
@@ -1,2 +1,3 @@
 veify.account.message.login=Check your email to verify your account!
-veify.account.message.expired=Yor verification link has expired. Please go to Create Account!
+veify.account.message.expired=Your verification link has expired. Please go to Create Account!
+veify.account.message.alredy.verified=Your account has been alredy verified. Please Login!
